Initialize transport service after performance() is called.
