+++ packages/performance/src/services/transport_service.ts
@@ -52,6 +52,10 @@ interface Log {
 
 let queue: BatchEvent[] = [];
 
+export function setupTransportService(): void {
+  processQueue(INITIAL_SEND_TIME_DELAY_MS);
+}
+
 function processQueue(timeOffset: number): void {
   setTimeout(() => {
     // If there is no remainingTries left, stop retrying.
@@ -120,8 +124,6 @@ function processQueue(timeOffset: number): void {
   }, timeOffset);
 }
 
-processQueue(INITIAL_SEND_TIME_DELAY_MS);
-
 function addToQueue(evt: BatchEvent): void {
   if (!evt.eventTime || !evt.message) {
     throw ERROR_FACTORY.create(ErrorCode.INVALID_CC_LOG);
@@ -131,7 +133,7 @@ function addToQueue(evt: BatchEvent): void {
 }
 
 /** Log handler for cc service to send the performance logs to the server. */
-export function ccHandler(
+export function transportHandler(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   serializer: (...args: any[]) => string
 ): (...args: unknown[]) => void {
